Moved event-related classes to their own namespace
