+++ src/ContentRelationshipMigrator.php
@@ -7,6 +7,7 @@ use Dhii\I18n\FormatTranslatorInterface;
 use Dhii\I18n\StringTranslatingTrait;
 use Dhii\I18n\StringTranslatorAwareTrait;
 use Inpsyde\MultilingualPress2to3\Db\DatabaseWpdbTrait;
+use Inpsyde\MultilingualPress2to3\Event\WpTriggerCapableTrait;
 use Throwable;
 use UnexpectedValueException;
 use wpdb as Wpdb;
