+++ src/MigrateCliCommandHandler.php
@@ -6,6 +6,7 @@ namespace Inpsyde\MultilingualPress2to3;
 use Dhii\I18n\StringTranslatingTrait;
 use Inpsyde\MultilingualPress2to3\Cli\AddCliCommandCapableWpTrait;
 use Inpsyde\MultilingualPress2to3\Config\ConfigAwareTrait;
+use Inpsyde\MultilingualPress2to3\Event\WpHookingTrait;
 use Inpsyde\MultilingualPress2to3\Handler\HandlerInterface;
 use Inpsyde\MultilingualPress2to3\Handler\HandlerTrait;
 use Psr\Container\ContainerInterface;
